Fix integration tests, refactor db package
