+++ packages/db/src/index.ts
@@ -1,16 +1,21 @@
-export { generateLocalTestDb, destroyLocalTestDb } from './test-utils';
 export {
-  createTestDatabase, deleteTestDatabase,
-} from './database';
+  generateLocalTestDb,
+  destroyLocalTestDb,
+  createTestDatabase,
+  deleteTestDatabase,
+} from './test-utils';
+
 export { getKnexClient } from './connection';
 export { getKnexConfig, localStackConnectionEnv } from './config';
 export { doesRecordExist, isRecordDefined } from './database';
 export { tableNames } from './tables';
+
 export {
   rdsProviderFromCumulusProvider,
   validateProviderHost,
   nullifyUndefinedProviderValues,
 } from './provider';
+
 export {
   AsyncOperationRecord,
   CollectionRecord,
